Product Admin - Securing the views

+++ products/views.py
@@ -3,6 +3,7 @@ from django.contrib import messages
 from .models import Product, Category
 from django.db.models import Q
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
 
 
 def all_products(request):
@@ -73,10 +74,16 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     Add a product to the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'You need to have the correct '
+                       'permissions to manage product details')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -97,10 +104,16 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     Edit a product in the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'You need to have the correct '
+                       'permissions to manage product details')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -124,10 +137,16 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     Delete a product from the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'You need to have the correct '
+                       'permissions to manage product details')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
